Pull shuffle() into file
